move cid to primitives
