+++ core/vm/actor/actor.hpp
@@ -8,9 +8,9 @@
 
 #include <libp2p/multi/content_identifier_codec.hpp>
 
-#include "common/cid.hpp"
 #include "primitives/address/address.hpp"
 #include "primitives/big_int.hpp"
+#include "primitives/cid/cid.hpp"
 
 namespace fc::vm::actor {
   using primitives::address::Address;
